kernels: move install functions to `install`
