+++ kernels/src/kernels/__init__.py
@@ -8,6 +8,7 @@ from kernels._windows import _add_additional_dll_paths
 from kernels.benchmark import Benchmark
 from kernels.hf_hub import RepoInfo
 from kernels.importer import LoadedKernel, get_loaded_kernels
+from kernels.install import install_kernel
 from kernels.layer import (
     CUDAProperties,
     Device,
@@ -32,7 +33,6 @@ from kernels.utils import (
     get_local_kernel,
     get_locked_kernel,
     has_kernel,
-    install_kernel,
     load_kernel,
 )
 from kernels.variants import (
